create domain package inside librrary and move model and transformer into it
