+++ ohs-player-library/src/commonMain/kotlin/dev/ohs/player/library/domain/transform/DataTransformer.kt
@@ -1,8 +1,8 @@
-package dev.ohs.player.library.transform
+package dev.ohs.player.library.domain.transform
 
 import dev.ohs.fhir.fhirpath.FhirPathEngine
 import dev.ohs.fhir.model.r4.Resource
-import dev.ohs.player.library.model.ViewDefinition
+import dev.ohs.player.library.domain.model.ViewDefinition
 import kotlinx.serialization.json.Json
