+++ ohs-player-library/src/commonMain/kotlin/dev/ohs/player/library/domain/transform/FhirExtractionResult.kt
@@ -1,4 +1,4 @@
-package dev.ohs.player.library.transform
+package dev.ohs.player.library.domain.transform
 
 import dev.ohs.fhir.model.r4.Resource
 import kotlinx.serialization.json.Json
